Split out imageLightbox.js from uploadPanel.js
